==> notebooks/00_problem_statement.ipynb <==
1️⃣ Problem Title
Agricultural Mandi Crop Price Prediction System

2️⃣ Real-World / Business Problem
    
     In India, crop prices vary daily across mandis due to demand, supply, weather, transportation, and market conditions.

     Farmers and traders often do not know future prices, which leads to:

        1. Selling crops at low prices
        2. Poor planning of storage and transportation
        3. Financial loss to farmers


    🎯 Goal
        To predict the future price of a crop in a specific mandi on a given date using historical mandi data.       

3️⃣ Dataset Description
    Each record represents crop prices in a mandi on a particular date.


    Sl no.: 1  
       District Name: Auraiya  
       Market Name: Achalda  
       Commodity: Wheat  
       Variety: Dara  
       Grade: FAQ  
       Min Price: 2350  
       Max Price: 2550  
       Modal Price: 2450  
       Price Date: 05-Apr-2025  
       State: Eg: Uttar Pradesh

4️⃣ Input Features (X)
The model will take the following inputs:

Feature Name_______Description
1. State	         State where mandi is located
2. District Name	 District of mandi
3. Market Name	     Specific mandi
4. Commodity	     Crop name (Wheat, Rice, Onion, etc.)
5. Variety	         Crop variety
6. Grade	         Crop quality grade
7. Min Price	     Minimum price on previous days
8. Max Price	     Maximum price on previous days
9. Price Date	     Date (converted to numerical features like day, month)

📌 Historical data (today, yesterday, previous days) will be used to predict future price.

5️⃣ Output (Target Variable – y)
🎯 Modal Price (Rs./Quintal)

This is the most frequently occurring price in the mandi and is the most reliable for farmers.

9 Example Prediction (Final Output Analysis)
Input:
State: Uttar Pradesh  
District: Auraiya  
Market: Achalda  
Commodity: Wheat  
Variety: Dara  
Grade: FAQ  
Date: 06-Apr-2025  
Previous Modal Prices: [2400, 2420, 2450]
Model Output:
Predicted Modal Price: ₹2485 / Quintal


**Interpretation:**
Price is expected to increase slightly

Farmer may choose to wait one more day before selling
