==> notebooks/03_data_cleaning.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../data/interim/ingested_data.csv")

df.shape

df.info()


df.shape

df.isna().sum()

df.shape

df.head()

df.columns = (
    df.columns
    .str.strip()
    .str.lower()
    .str.replace(" ", "_")
    .str.replace(r"[()/.]", "", regex=True)
)


df.head()

df[["min_price_rsquintal", "max_price_rsquintal", "modal_price_rsquintal"]].dtypes


df.shape

df["price_date"] = pd.to_datetime(df["price_date"], errors="coerce")
df["price_date"].dtype

df.head()

df["modal_price_rsquintal"].isna().sum()

df = df.dropna(subset=["modal_price_rsquintal"])
df["modal_price_rsquintal"].shape

df.shape

df = df.dropna(subset=["price_date"])
df["price_date"].shape


df.shape

df = df.rename(columns={
    "sl_no": "id",
    "district_name": "district",
    "market_name": "market",
    "min_price_rsquintal": "min_price",
    "max_price_rsquintal": "max_price",
    "modal_price_rsquintal": "modal_price",
    "price_date": "date"
})
df.head()

df.sample(5)

df[df["min_price"].isna() | df["max_price"].isna()]


df[df["modal_price"].isna()]


df.loc[
    df[["min_price", "max_price", "modal_price"]].isna().any(axis=1)
]


df.shape

df[
    (df["min_price"] <= 0) |
    (df["max_price"] <= 0) |
    (df["modal_price"] <= 0)].shape

df = df[
    (df["min_price"] > 0) &
    (df["max_price"] > 0) &
    (df["modal_price"] > 0)
]
df.shape


df = df[df["min_price"] <= df["max_price"]]


df = df[
    (df["modal_price"] >= df["min_price"]) &
    (df["modal_price"] <= df["max_price"])
]


df.shape

df = df.dropna(subset=["state", "district", "market", "commodity"])


df.shape

df = df.drop_duplicates()
df.shape

duplicate_rows = df[
    df.duplicated(
        subset=["market", "commodity", "date", "grade", "variety", "state", "district"],
        keep=False
    )
]


duplicate_rows

🎯 One-line rule (MEMORIZE THIS)
    A mandi is defined by (state + district + market), not market name alone.

gourped_rows = duplicate_rows.groupby(
    ["market", "commodity", "date", "grade", "variety", "state", "district"]
).ngroups


df[
    (df["market"] == "Gorakhpur") &
    (df["commodity"] == "Maize") &
    (df["date"] == "2024-11-28")&
    (df["grade"] == "FAQ") &
    (df["state"] == "Uttar Pradesh") &
    (df["district"] == "Gorakhpur")
]


df.shape

df.drop_duplicates(
    subset=["market", "commodity", "date", "grade", "variety", "state", "district"],
    keep="first"
)
df.shape


df.groupby(["market", "commodity", "date", "grade", "variety", "state", "district"]).size().reset_index(name="count") \
  .query("count > 1") \
  .head(10)

df.head(5)

# With these steps cleaning of data is complete now 

df.to_csv("../data/processed/cleaned_data.csv", index=False)
